+++ stacks-signer/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 ### Changed
 
 - Add tenure extend timestamp to signer block responses
+- Added tenure_idle_timeout_secs configuration option for determining when a tenure extend will be accepted
 
 ## [3.0.0.0.1.0]